% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biRF.R
\name{biRF}
\alias{biRF}
\title{Balanced iterative random forest}
\usage{
biRF(dat, group, ...)
}
\arguments{
\item{dat}{matrix of features in columns and samples in rows}

\item{group}{vector of classifications for each sample in the same order as the rows in the matrix}

\item{...}{all other arguments/parameters for the randomForest function from the randomForest package}
}
\description{
A random forest embedded method for feature selection
}
\examples{
biRF()
}
\keyword{balanced}
\keyword{feature}
\keyword{forest}
\keyword{random}
\keyword{selection}
