% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaTable.R
\name{limmaTable}
\alias{limmaTable}
\title{A summary function for limma fits}
\usage{
limmaTable(dat, p, fit, grouping, contrast.matrix, adj.method, pval)
}
\arguments{
\item{dat}{a matrix of the values of interest with the corresponding sample labels and probe labels}

\item{p}{a dataframe containing the group column matched with the sample labels in dat}

\item{fit}{a limma fit}

\item{grouping}{the name of the column in which has the group parameter}

\item{contrast.matrix}{a model matrix; make sure the levels of the contrasts are the same as the levels in the grouping variable}

\item{adj.method}{the multiple testings correction method as used in topTable function, i.e. fdr BH etc.}

\item{pval}{the desired p-value}
}
\description{
Organizing p-values and methylation differences between groups into a single data frame
}
\examples{
limmaTable()
}
\keyword{microarray}
