% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.anova.R
\name{pairwise.anova}
\alias{pairwise.anova}
\title{A pairwise testing using tukey HSD after anova modelling}
\usage{
pairwise.anova(dat, var, grouping)
}
\arguments{
\item{dat}{a dataframe with at least a column with the grouping and another with the numeric data}

\item{var}{the name of the column of the dependent variable}

\item{grouping}{the name of the column of the independing variable, i.e. the grouping}
}
\description{
This function allows you to do tukey pairwise testings of numeric data between members of a population with >2 groups after modelling with anova
}
\examples{
pairwise.anova()
}
\keyword{pairwise}
