% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.fisher.R
\name{pairwise.fisher}
\alias{pairwise.fisher}
\title{A pairwise testing using fisher}
\usage{
pairwise.fisher(dat, var, grouping, adj.method)
}
\arguments{
\item{dat}{a dataframe containing at least a column with the grouping and another with the categorical value}

\item{var}{the name of the column of the dependent variable}

\item{grouping}{the name of the column of the independent variable, i.e. the grouping of interest}

\item{adj.method}{the multiple testings correction method as used in p.adjust, e.g. "bonferroni" or "BH" or "fdr"}
}
\description{
This function allows you to do fisher.test pairwise testings of categorical count data between members of a population with >2 groups
}
\examples{
pairwise.fisher()
}
\keyword{pairwise}
