% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.wilcox.R
\name{pairwise.wilcox}
\alias{pairwise.wilcox}
\title{A pairwise testing using wilcox}
\usage{
pairwise.wilcox(dat, var, grouping, adj.method)
}
\arguments{
\item{dat}{a dataframe with at least a column with the grouping and another with the numeric data}

\item{var}{the name of the column of the dependent variable}

\item{grouping}{the name of the column of the independing variable, i.e. the grouping}

\item{adj.method}{the multiple testings correction method as used in p.adjust, e.g. "bonferroni" or "BH" or "fdr"}
}
\description{
This function allows you to do wilcox pairwise testings of numeric data between members of a population with >2 groups
}
\examples{
pairwise.wilcox()
}
\keyword{pairwise}
