% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.cov.cor.R
\name{pca.cov.cor}
\alias{pca.cov.cor}
\title{Mining PCA algorithm}
\usage{
pca.cov.cor(dat, p, ignore, adjustment, showall)
}
\arguments{
\item{dat}{matrix with desired measure to run pca on}

\item{p}{dataframe of covariates to test with each pc; covariate data should match with the pca order of samples}

\item{ignore}{columns to ignore or not use}

\item{adjustment}{the method of multiple adjustment for iterative testing of covariates within each PC, e.g. bonferroni, same arguments for p.adjust}

\item{showall}{TRUE or FALSE; if TRUE, a matrix of all pvalues will be shown}
}
\description{
pca algorithm to iteratively check principal componenets with matched sample covariate data
}
\examples{
pca.cov()
}
\keyword{microarray}
