% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.cov.R
\name{pca.cov}
\alias{pca.cov}
\title{Mining PCA algorithm}
\usage{
pca.cov(dat, p, ignore, showall)
}
\arguments{
\item{dat}{matrix with desired measure to run pca on}

\item{p}{dataframe of covariates to test with each pc; covariate data should match with the pca order of samples; covariate data should be categorical}

\item{ignore}{columns to ignore or not use}

\item{showall}{TRUE or FALSE; if TRUE, a matrix of all pvalues will be shown}
}
\description{
pca algorithm to iteratively check principal componenets with matched sample covariate data
}
\examples{
pca.cov()
}
\keyword{microarray}
