% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaTopCandidates.R
\name{pcaTopCandidates}
\alias{pcaTopCandidates}
\title{Determination of top weighted eigengenes for a specific PC}
\usage{
pcaTopCandidates(dat, PCnumber)
}
\arguments{
\item{dat}{matrix with desired measure to run pca on}

\item{PCnumber}{the PC of interest}
}
\description{
function to extract out the top weighted eigengenes of a given matrix and specific PC; top is determined by 1.5* the IQR of the eigenvector
}
\examples{
pcaTopCandidates()
}
\keyword{microarray}
