% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfLoocv.R
\name{rfLoocv}
\alias{rfLoocv}
\title{Leave-one-out-cross-validation with a random forest classifier}
\usage{
rfLoocv(dat, group, ...)
}
\arguments{
\item{dat}{matrix of features in columns and samples in rows}

\item{group}{vector of classifications for each sample in the same order as the rows in the matrix}

\item{...}{all other arguments/parameters for the randomForest function from the randomForest package}
}
\description{
a cross validation method to check the accuracy of a svm classifier on a matrix of features and samples
}
\examples{
rfLoocv()
}
\keyword{forest}
\keyword{leave-one-out-cross-validation}
\keyword{random}
