% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.s.test.R
\name{s.s.test}
\alias{s.s.test}
\title{Summary statistics by group}
\usage{
s.s.test(dat, samplelabels, ignore)
}
\arguments{
\item{dat}{dataframe with desired summary statistics to be calculated}

\item{samplelabels}{the name of the column in which has the group parameter you are interested in}

\item{ignore}{columns to ignore or not use}
}
\description{
Summary and statistical comparison of group characteristics
}
\examples{
s.s.test()
}
\keyword{microarray}
