% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmLoocv.R
\name{svmLoocv}
\alias{svmLoocv}
\title{Leave-one-out-cross-validation with a Support Vector Machine classifier}
\usage{
svmLoocv(dat, group, ...)
}
\arguments{
\item{dat}{matrix of features in columns and samples in rows}

\item{group}{vector of classifications for each sample in the same order as the rows in the matrix}

\item{...}{all other arguments/parameters for the svm function from the e1071 package}
}
\description{
a cross validation method to check the accuracy of a svm classifier on a matrix of features and samples
}
\examples{
svmLoocv()
}
\keyword{leave-one-out-cross-validation}
\keyword{svm}
