% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmRFE.R
\name{svmRFE}
\alias{svmRFE}
\title{Support vector machine recursive feature elimination 'svm-rfe'}
\usage{
svmRFE(dat, group)
}
\arguments{
\item{dat}{a matrix of n samples x m features}

\item{group}{a class of factor of the coresponding classes for each of the n samples in the matrix}
}
\description{
This function allows you to conduct recursive feature elimination using support vector machines
}
\examples{
svmRFE
}
\keyword{learning}
\keyword{machine}
\keyword{rfe}
\keyword{svm}
